refactor(cli): move logger and reporter into lib for better reuse
Centralize logger and reporter implementations under lib.
