+++ packages/cli/src/commands/reports/prune/index.ts
@@ -1,5 +1,5 @@
 import { resolveCommandPath } from '../../../utils/filesystem';
-import { Reporter } from '../../../utils/reporter';
+import { Reporter } from '../../../lib/reporter/reporter';
 import { getUI } from '../../../utils/ui';
 import { reportsCommand } from '../command';
 import { directories } from '../../../constants';
