+++ packages/cli/src/lib/reporter/reporter.ts
@@ -1,4 +1,4 @@
-import { saveToFileSync } from './filesystem';
+import { saveToFileSync } from '../../utils/filesystem';
 import { existsSync, readdirSync, unlinkSync } from 'node:fs';
 import { dirname, extname, join, relative } from 'node:path';
